Extracted Parser (Calculator)
